Crate sorted_vec

Sorted vectors.

Repository

SortedVec – sorted from least to greatest, may contain duplicates

SortedSet – sorted from least to greatest, unique elements

ReverseSortedVec – sorted from greatest to least, may contain duplicates

ReverseSortedSet – sorted from greatest to least, unique elements

The partial module provides sorted vectors of types that only implement PartialOrd where comparison of incomparable elements results in runtime panic.

Modules

partial

Sorted vectors of types implementing PartialOrd.

Structs

SortedSet

Forward sorted set

SortedVec

Forward sorted vector

Enums

FindOrInsert

Value returned when find_or_insert is used.

Type Aliases

ReverseSortedSet

ReverseSortedVec

Reverse-sorted Containers.