Struct sorted_vec::SortedVec

pub struct SortedVec<T> { /* fields omitted */ }

A set based on a 2-level rotated array.

See [README.md] for a detailed discussion of this collection's performance benefits and drawbacks.

Examples

use sorted_vec::SortedVec;

// Type inference lets us omit an explicit type signature (which
// would be `SortedVec<i32>` in this example).
let mut ints = SortedVec::new();

// Add some integers.
ints.insert(-1);
ints.insert(6);
ints.insert(1729);
ints.insert(24);

// Check for a specific one.
if !ints.contains(&42) {
    println!("We don't have the answer to Life, the Universe, and Everything.");
}

// Remove an integer.
ints.remove(&6);

// Iterate over everything.
for int in &ints {
    println!("{}", int);
}

Methods

impl<T> SortedVec<T> where
T: Ord + Copy + Default + Debug,

pub fn new() -> Self

Makes a new SortedVec without any heap allocations.

This is a constant-time operation.

Examples

use sorted_vec::SortedVec;

let mut set: SortedVec<i32> = SortedVec::new();

pub fn clear(&mut self)

Clears the set, removing all values.

This is a constant-time operation.

Examples

use sorted_vec::SortedVec;

let mut v = SortedVec::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

pub fn contains(&self, value: &T) -> bool

Returns true if the set contains a value.

This is an O(lg n) operation.

Examples

use sorted_vec::SortedVec;

let set: SortedVec<_> = vec![1, 2, 3].into();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

pub fn get(&self, value: &T) -> Option<&T>

Returns a reference to the value in the set, if any, that is equal to the given value.

This is an O(lg n) operation.

Examples

use sorted_vec::SortedVec;

let set: SortedVec<_> = vec![1, 2, 3].into();
assert_eq!(set.get(&2), Some(&2));
assert_eq!(set.get(&4), None);

pub fn at(&self, rank: usize) -> Option<&T>

Returns a reference to the value in the set, if any, with the given rank.

This is a constant-time operation.

Examples

use sorted_vec::SortedVec;

let set: SortedVec<_> = vec![1, 2, 3].into();
assert_eq!(set.at(0), Some(&1));
assert_eq!(set.at(3), None);

pub fn insert(&mut self, value: T) -> bool

Adds a value to the set.

This is an O(√n) operation.

If the set did not have this value present, true is returned.

If the set did have this value present, false is returned, and the entry is not updated.

Examples

use sorted_vec::SortedVec;

let mut set = SortedVec::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

pub fn remove(&mut self, value: &T) -> bool

Removes a value from the set. Returns whether the value was present in the set.

This is an O(√n) operation.

Examples

use sorted_vec::SortedVec;

let mut set = SortedVec::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);

pub fn take(&mut self, value: &T) -> Option<T>

Removes and returns the value in the set, if any, that is equal to the given one.

This is an O(√n) operation.

Examples

use sorted_vec::SortedVec;

let mut set: SortedVec<_> = vec![1, 2, 3].into();
assert_eq!(set.take(&2), Some(2));
assert_eq!(set.take(&2), None);

pub fn len(&self) -> usize

Returns the number of elements in the set.

This is a constant-time operation.

Examples

use sorted_vec::SortedVec;

let mut v = SortedVec::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

This is a constant-time operation.

Examples

use sorted_vec::SortedVec;

let mut v = SortedVec::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> where
T: Ord + Copy + Default + Debug, type Item = &'a T;

pub fn iter(&self) -> Iter<T>

Gets an iterator that visits the values in the SortedVec in ascending order.

Examples

use sorted_vec::SortedVec;

let set: SortedVec<usize> = vec![1, 2, 3].into();
let mut set_iter = set.iter();
assert_eq!(set_iter.next(), Some(&1));
assert_eq!(set_iter.next(), Some(&2));
assert_eq!(set_iter.next(), Some(&3));
assert_eq!(set_iter.next(), None);

Values returned by the iterator are returned in ascending order:

use sorted_vec::SortedVec;

let set: SortedVec<usize> = vec![3, 1, 2].into();
let mut set_iter = set.iter();
assert_eq!(set_iter.next(), Some(&1));
assert_eq!(set_iter.next(), Some(&2));
assert_eq!(set_iter.next(), Some(&3));
assert_eq!(set_iter.next(), None);