Crate sorted_vec

A set based on a 2-level rotated array.

See the repository README for a detailed discussion of this collection's performance benefits and drawbacks.

Examples

use sorted_vec::SortedVec;

// Type inference lets us omit an explicit type signature (which
// would be `SortedVec<i32>` in this example).
let mut ints = SortedVec::new();

// Add some integers.
ints.insert(-1);
ints.insert(6);
ints.insert(1729);
ints.insert(24);

// Check for a specific one.
if !ints.contains(&42) {
    println!("We don't have the answer to Life, the Universe, and Everything :-(");
}

// Remove an integer.
ints.remove(&6);

// Iterate over everything.
for int in &ints {
    println!("{}", int);
}

Structs

Difference
A lazy iterator producing elements in the difference of SortedVecs.

Intersection
A lazy iterator producing elements in the intersection of SortedVecs.

IntoIter
An owning iterator over the items of a SortedVec.

Iter
An iterator over the items of a SortedVec.

SortedVec
A sorted set data structure, containing a data array and two auxiliary arrays respectively listing the index of the minimum element in each chunk, and the minimum element itself, for efficient binary search.

SymmetricDifference
A lazy iterator producing elements in the symmetric difference of SortedVecs.

Union
A lazy iterator producing elements in the union of SortedVecs.